remove user profile class, use player class instead several minor fixes
