+++ app/src/main/res/layout/activity_practice.xml
@@ -158,7 +158,7 @@
         android:id="@+id/correct_answer_count_label"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="10"
+        android:text="0"
         android:textAlignment="center"
         android:textColor="@color/green"
         android:textSize="40sp"
@@ -170,7 +170,7 @@
         android:id="@+id/question_count_label"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="10"
+        android:text="0"
         android:textAlignment="center"
         android:textColor="@color/colorPrimary"
         android:textSize="40sp"
